+++ src/Service/MicrosoftGraphHelperService.php
@@ -1,0 +1,82 @@
+<?php
+
+namespace App\Service;
+
+use App\Exception\MicrosoftGraphCommunicationException;
+use App\Factory\ClientFactory;
+use GuzzleHttp\Exception\GuzzleException;
+use JsonException;
+use Microsoft\Graph\Exception\GraphException;
+use Microsoft\Graph\Http\GraphResponse;
+use Psr\Cache\CacheItemInterface;
+use Symfony\Contracts\Cache\CacheInterface;
+
+class MicrosoftGraphHelperService
+{
+    public function __construct(
+        private readonly string $tenantId,
+        private readonly string $clientId,
+        private readonly string $serviceAccountUsername,
+        private readonly string $serviceAccountPassword,
+        private readonly CacheInterface $graphCache,
+        private readonly ClientFactory $clientFactory,
+    ) {
+    }
+
+    public function authenticateAsServiceAccount(): string
+    {
+        return $this->graphCache->get('serviceAccountToken', function (CacheItemInterface $item) {
+            $tokenEntry = $this->authenticateAsUser($this->serviceAccountUsername, $this->serviceAccountPassword);
+
+            $item->expiresAfter($tokenEntry['expires_in'] ?? 3600);
+
+            return $tokenEntry['access_token'];
+        });
+    }
+
+    /**
+     * @throws MicrosoftGraphCommunicationException
+     */
+    public function authenticateAsUser(string $username, string $password): array
+    {
+        try {
+            $client = $this->clientFactory->getGuzzleClient();
+            $url = 'https://login.microsoftonline.com/'.$this->tenantId.'/oauth2/v2.0/token';
+
+            $response = $client->post($url, [
+                'form_params' => [
+                    'client_id' => $this->clientId,
+                    'scope' => 'https://graph.microsoft.com/.default',
+                    'username' => $username,
+                    'password' => $password,
+                    'grant_type' => 'password',
+                ],
+            ]);
+
+            return json_decode($response->getBody()->getContents(), true, 512, JSON_THROW_ON_ERROR);
+        } catch (JsonException|GuzzleException $exception) {
+            throw new MicrosoftGraphCommunicationException($exception->getMessage(), $exception->getCode());
+        }
+    }
+
+    /**
+     * @throws MicrosoftGraphCommunicationException
+     */
+    public function request(string $path, string $accessToken, string $requestType = 'GET', array $body = null): GraphResponse
+    {
+        try {
+            $graph = $this->clientFactory->getGraph();
+            $graph->setAccessToken($accessToken);
+
+            $graphRequest = $graph->createRequest($requestType, $path);
+
+            if ($body) {
+                $graphRequest->attachBody($body);
+            }
+
+            return $graphRequest->execute();
+        } catch (GuzzleException|GraphException $e) {
+            throw new MicrosoftGraphCommunicationException($e->getMessage(), $e->getCode());
+        }
+    }
+}