+++ src/Service/NotificationService.php
@@ -7,9 +7,10 @@ use App\Entity\Main\UserBooking;
 use App\Entity\Resources\AAKResource;
 use App\Enum\NotificationTypeEnum;
 use App\Utils\ValidationUtils;
-use DateTimeImmutable;
+use DateTimeZone as PhpDateTimeZone;
 use Eluceo\iCal\Domain\Entity;
-use Eluceo\iCal\Domain\ValueObject\DateTime;
+use Eluceo\iCal\Domain\Entity\TimeZone;
+use Eluceo\iCal\Domain\ValueObject\DateTime as ICalDateTime;
 use Eluceo\iCal\Domain\ValueObject\GeographicPosition;
 use Eluceo\iCal\Domain\ValueObject\Location;
 use Eluceo\iCal\Domain\ValueObject\TimeSpan;
@@ -33,7 +34,9 @@ class NotificationService implements NotificationServiceInterface
         private readonly string $emailAdminNotification,
         private readonly ValidationUtils $validationUtils,
         private readonly LoggerInterface $logger,
-        private readonly MailerInterface $mailer
+        private readonly MailerInterface $mailer,
+        private readonly string $bindNotificationTimezone,
+        private readonly string $bindNotificationDateFormat,
     ) {
         try {
             $this->validatedAdminNotificationEmail = $this->validationUtils->validateEmail($this->emailAdminNotification);
@@ -91,6 +94,12 @@ class NotificationService implements NotificationServiceInterface
 
         $name = $node->nodeValue;
 
+        $dateStart = $userBooking->start;
+        $dateEnd = $userBooking->end;
+
+        $dateStart->setTimezone(new \DateTimeZone($this->bindNotificationTimezone));
+        $dateEnd->setTimezone(new \DateTimeZone($this->bindNotificationTimezone));
+
         $notificationData = [
             'from' => $this->emailFromAddress,
             'to' => $email,
@@ -110,11 +119,11 @@ class NotificationService implements NotificationServiceInterface
                 'resource' => [
                     'resourceName' => $userBooking->resourceName,
                 ],
+                'startFormatted' => $dateStart->format($this->bindNotificationDateFormat),
+                'endFormatted' => $dateEnd->format($this->bindNotificationDateFormat),
             ],
             'fileAttachments' => [],
         ];
-
-        $notifyResourceSubject = null;
 
         switch ($type) {
             case NotificationTypeEnum::DELETE_SUCCESS:
@@ -132,6 +141,8 @@ class NotificationService implements NotificationServiceInterface
                 break;
             default:
                 $this->logger->error('Error sending UserBooking notification: Unsupported NotificationTypeEnum');
+
+                return;
         }
 
         $this->sendNotification($notificationData);
@@ -148,45 +159,44 @@ class NotificationService implements NotificationServiceInterface
     /**
      * {@inheritdoc}
      */
-    public function createCalendarComponent(array $events): Component
+    public function createCalendarComponent(array $eventData): Component
     {
-        $iCalEvents = [];
+        $location = new Location($eventData['location_name']);
 
-        foreach ($events as $eventData) {
-            $location = new Location($eventData['location_name']);
-
-            if ($eventData['coordinates']) {
-                $coordinatesArr = explode(',', $eventData['coordinates']);
-                $location = $location->withGeographicPosition(
-                    new GeographicPosition(
-                        (float) $coordinatesArr['0'],
-                        (float) $coordinatesArr['1']
-                    )
-                );
-            }
-
-            $event = new Entity\Event();
-
-            $event->setSummary($eventData['summary']);
-            $event->setDescription($eventData['description']);
-
-            $immutableFrom = DateTimeImmutable::createFromFormat('Y-m-d H:i:s', $eventData['from']);
-            $immutableTo = DateTimeImmutable::createFromFormat('Y-m-d H:i:s', $eventData['from']);
-
-            if (false === $immutableFrom || false === $immutableTo) {
-                throw new Exception('DateTimeImmutable cannot be false');
-            }
-
-            $start = new DateTime($immutableFrom, false);
-            $end = new DateTime($immutableTo, false);
-            $occurrence = new TimeSpan($start, $end);
-            $event->setOccurrence($occurrence);
-            $event->setLocation($location);
-
-            $iCalEvents[] = $event;
+        if ($eventData['coordinates']) {
+            $coordinatesArr = explode(',', $eventData['coordinates']);
+            $location = $location->withGeographicPosition(
+                new GeographicPosition(
+                    (float) $coordinatesArr['0'],
+                    (float) $coordinatesArr['1']
+                )
+            );
         }
 
-        $calendar = new Entity\Calendar($iCalEvents);
+        $event = new Entity\Event();
+
+        $event->setSummary($eventData['summary']);
+        $event->setDescription($eventData['description']);
+        $event->setLocation($location);
+
+        $dateFrom = $eventData['start']->setTimezone(new \DateTimeZone($this->bindNotificationTimezone));
+        $dateTo = $eventData['end']->setTimezone(new \DateTimeZone($this->bindNotificationTimezone));
+
+        $start = new ICalDateTime($dateFrom, true);
+        $end = new ICalDateTime($dateTo, true);
+
+        $occurrence = new TimeSpan($start, $end);
+        $event->setOccurrence($occurrence);
+
+        $calendar = new Entity\Calendar([$event]);
+
+        $phpDateTimeZone = new PhpDateTimeZone($this->bindNotificationTimezone);
+        $timeZone = TimeZone::createFromPhpDateTimeZone(
+            $phpDateTimeZone,
+            $dateFrom,
+            $dateTo
+        );
+        $calendar->addTimeZone($timeZone);
 
         return (new Factory\CalendarFactory())->createCalendar($calendar);
     }
@@ -200,11 +210,24 @@ class NotificationService implements NotificationServiceInterface
             $to = $this->validatedAdminNotificationEmail;
             $template = 'email-notify-admin.html.twig';
 
+            if (null !== $booking) {
+                $dateStart = $booking->getStartTime();
+                $dateEnd = $booking->getEndTime();
+
+                $dateStart->setTimezone(new \DateTimeZone($this->bindNotificationTimezone));
+                $dateEnd->setTimezone(new \DateTimeZone($this->bindNotificationTimezone));
+
+                $dateStartString = $dateStart->format($this->bindNotificationDateFormat);
+                $dateEndString = $dateEnd->format($this->bindNotificationDateFormat);
+            }
+
             $data = [
                 'subject' => $subject,
                 'message' => $message,
                 'booking' => $booking,
                 'resource' => $resource,
+                'startFormatted' => $dateStartString ?? '',
+                'endFormatted' => $dateEndString ?? '',
             ];
 
             $notificationData = [
@@ -230,8 +253,6 @@ class NotificationService implements NotificationServiceInterface
         $notificationData = [];
 
         try {
-            $template = null;
-            $subject = null;
             $fileAttachments = [];
             $to = $data['user']['mail'];
 
@@ -240,8 +261,8 @@ class NotificationService implements NotificationServiceInterface
                     $template = 'email-booking-success.html.twig';
                     $subject = 'Booking bekræftelse: '.$data['resource']->getResourceName().' - '.$data['resource']->getLocation();
 
-                    $events = $this->prepareIcalEvents($data);
-                    $iCalendarComponent = $this->createCalendarComponent($events);
+                    $event = $this->prepareICalEvent($data);
+                    $iCalendarComponent = $this->createCalendarComponent($event);
 
                     $fileAttachments = [
                         'ics' => [$iCalendarComponent],
@@ -259,6 +280,22 @@ class NotificationService implements NotificationServiceInterface
                     $template = 'email-booking-failed.html.twig';
                     $subject = 'Booking lykkedes ikke. Intervallet er optaget: '.$data['resource']->getResourceName().' - '.$data['resource']->getLocation();
                     break;
+                default:
+                    throw new Exception('Unsupported NotificationTypeEnum');
+            }
+
+            if (isset($data['booking'])) {
+                /** @var Booking $booking */
+                $booking = $data['booking'];
+
+                $dateStart = $booking->getStartTime();
+                $dateEnd = $booking->getEndTime();
+
+                $dateStart->setTimezone(new \DateTimeZone($this->bindNotificationTimezone));
+                $dateEnd->setTimezone(new \DateTimeZone($this->bindNotificationTimezone));
+
+                $data['startFormatted'] = $dateStart->format($this->bindNotificationDateFormat);
+                $data['endFormatted'] = $dateEnd->format($this->bindNotificationDateFormat);
             }
 
             $notificationData = [
@@ -323,19 +360,17 @@ class NotificationService implements NotificationServiceInterface
     /**
      * @param array $data
      *
-     * @return array[]
+     * @return array
      */
-    private function prepareIcalEvents(array $data): array
+    private function prepareICalEvent(array $data): array
     {
         return [
-            [
-                'summary' => $data['booking']->getSubject(),
-                'description' => $data['booking']->getBody(),
-                'from' => $data['booking']->getStartTime()->format('Y-m-d H:i:s'),
-                'to' => $data['booking']->getEndTime()->format('Y-m-d H:i:s'),
-                'coordinates' => $data['resource']->getGeoCoordinates(),
-                'location_name' => $data['resource']->getLocation().' - '.$data['resource']->getResourceName(),
-            ],
+            'summary' => $data['booking']->getSubject(),
+            'start' => $data['booking']->getStartTime(),
+            'end' => $data['booking']->getEndTime(),
+            'description' => $data['booking']->getSubject(),
+            'coordinates' => $data['resource']->getGeoCoordinates(),
+            'location_name' => $data['resource']->getLocation().' - '.$data['resource']->getResourceName(),
         ];
     }
 }